+++ examples/Basic/1-Widom/run_widom.py
@@ -1,0 +1,58 @@
+import os
+
+# Hide UserWarning and RuntimeWarning messages
+import warnings
+
+import ase
+import torch
+from ase.data import vdw_radii
+from ase.io import read
+from mace.calculators import mace_mp
+
+from mlp_adsorption.widom import Widom
+
+warnings.filterwarnings("ignore", category=UserWarning)
+warnings.filterwarnings("ignore", category=RuntimeWarning)
+
+os.environ["KMP_DUPLICATE_LIB_OK"] = "TRUE"
+
+device = "cuda" if torch.cuda.is_available() else "cpu"
+
+FrameworkPath = "mg-mof-74.cif"
+AdsorbatePath = "co2.xyz"
+
+model = mace_mp(
+    model="medium-0b2",
+    dispersion=True,
+    damping="zero",  # choices: ["zero", "bj", "zerom", "bjm"]
+    dispersion_xc="pbe",
+    default_dtype="float32",
+    device=device,
+)
+
+
+# Load the framework structure
+framework: ase.Atoms = read(FrameworkPath)  # type: ignore
+
+# Load the adsorbate structure
+adsorbate: ase.Atoms = read(AdsorbatePath)  # type: ignore
+
+Temperature = 298.0
+
+NSteps = 30000
+
+widom = Widom(
+    model=model,
+    framework_atoms=framework,
+    adsorbate_atoms=adsorbate,
+    temperature=Temperature,
+    device=device,
+    vdw_radii=vdw_radii,
+    debug=False,
+    output_to_file=True,
+)
+
+widom.print_header()
+
+widom.run(NSteps)
+widom.print_footer()